#12 [Enum & Interface Updates]
