+++ meda-mailbox/src/app/common/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './page-path.enum';
+export * from './mailbox.enum';
