+++ meda-mailbox/src/app/common/enums/mailbox.enum.ts
@@ -1,0 +1,4 @@
+export enum MailboxViewEnum {
+  SPLIT = 'split',
+  PANEL = 'panel',
+}
