+++ meda-mailbox/src/app/common/interfaces/index.ts
@@ -1,3 +1,2 @@
-export * from './store.interface';
 export * from './mailbox.interface';
 export * from './component.interface';
