+++ meda-mailbox/src/app/common/interfaces/mailbox.interface.ts
@@ -1,3 +1,12 @@
+export interface MailboxInterface {
+  total: number;
+  unread: number;
+  message?: any;
+  messages: any[];
+  fetchingMessages?: boolean,
+  currentView: MailboxViewType,
+}
+
 export interface MailboxMessagesInterface {
   id: string,
   read:boolean,
@@ -5,3 +14,5 @@ export interface MailboxMessagesInterface {
   preview: string,
   content?: string
 }
+
+export type MailboxViewType = 'split' | 'panel';
