--- meda-mailbox/src/app/common/interfaces/store.interface.ts
@@ -1,7 +1,0 @@
-export interface MailboxInterface {
-  total: number;
-  unread: number;
-  message?: any;
-  messages: any[];
-  fetchingMessages?: boolean,
-}
